+++ src/kalman.cpp
@@ -2,7 +2,7 @@
 
 namespace keisan {
 
-Kalman::Kalman(float dt, float std_dev_aceleration,
+Kalman::Kalman(double dt, double std_dev_aceleration,
                Matrix<2, 1> std_measurement, Matrix<2, 1> acceleration)
     : U(acceleration) {
   Xk = Matrix<4, 1>::zero();
